requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
filelock==3.0.12
idna==2.10
joblib==1.0.1
lxml==4.6.3
numpy==1.20.3
packaging==20.9
Pillow==8.2.0
praw==7.2.0
prawcore==2.0.0
pyparsing==2.4.7
python-dotenv==0.17.1
regex==2021.4.4
requests==2.25.1
sacremoses==0.0.45
six==1.16.0
soupsieve==2.2.1
tokenizers==0.10.2
torch==1.8.1
torchaudio==0.8.1
torchvision==0.9.1
tqdm==4.60.0
transformers==4.5.1
typing-extensions==3.10.0.0
update-checker==0.18.0
urllib3==1.26.4
websocket-client==0.59.0
